pass correct variable to nvim_buf_get_option
Previously we were passing 'buf', which is undefined, evaluating to nil.
It looks like in previous versions of neovim, this was accepted
(probably coerced to the number 0), and evaluated as the current buffer.
At some point, this was changed and newer versions of neovim will throw
an error because the argument is not a number.